latest change

+++ app/models/orders.model.js
@@ -40,97 +40,97 @@ module.exports = ( sequelize, Sequelize, DataTypes ) =>
                 field: 'transaction_id'
             },
             bill_amount: {
-                type: DataTypes.DECIMAL,
+                type: DataTypes.FLOAT( 10, 2 ),
                 allowNull: true,
                 field: 'bill_amount',
                 defaultValue: 0.00
             },
             discount_from_restaurant: {
-                type: DataTypes.DECIMAL,
+                type: DataTypes.FLOAT( 10, 2 ),
                 allowNull: true,
                 field: 'discount_from_restaurant',
                 defaultValue: 0.00
             },
             discount_to_customer: {
-                type: DataTypes.DECIMAL,
+                type: DataTypes.FLOAT( 10, 2 ),
                 allowNull: true,
                 field: 'discount_to_customer',
                 defaultValue: 0.00
             },
             discount_commision: {
-                type: DataTypes.DECIMAL,
+                type: DataTypes.FLOAT( 10, 2 ),
                 allowNull: true,
                 field: 'discount_commision',
                 defaultValue: 0.00
             },
             magic_coupon_discount: {
-                type: DataTypes.DECIMAL,
+                type: DataTypes.FLOAT( 10, 2 ),
                 allowNull: true,
                 field: 'magic_coupon_discount',
                 defaultValue: 0.00
             },
             convinence_fee: {
-                type: DataTypes.DECIMAL,
+                type: DataTypes.FLOAT( 10, 2 ),
                 allowNull: true,
                 field: 'convinence_fee',
                 defaultValue: 0.00
             },
             gst: {
-                type: DataTypes.DECIMAL,
+                type: DataTypes.FLOAT( 10, 2 ),
                 allowNull: true,
                 field: 'gst',
                 defaultValue: 0.00
             },
             dis_to_customer: {
-                type: DataTypes.DECIMAL,
+                type: DataTypes.FLOAT( 10, 2 ),
                 allowNull: true,
                 field: 'dis_to_customer',
                 defaultValue: 0.00
             },
             amt_pay_by_customer: {
-                type: DataTypes.DECIMAL,
+                type: DataTypes.FLOAT( 10, 2 ),
                 allowNull: true,
                 field: 'amt_pay_by_customer',
                 defaultValue: 0.00
             },
             dis_receive_by_res: {
-                type: DataTypes.DECIMAL,
+                type: DataTypes.FLOAT( 10, 2 ),
                 allowNull: true,
                 field: 'dis_receive_by_res',
                 defaultValue: 0.00
             },
             commission_by_admin: {
-                type: DataTypes.DECIMAL,
+                type: DataTypes.FLOAT( 10, 2 ),
                 allowNull: true,
                 field: 'commission_by_admin',
                 defaultValue: 0.00
             },
             magic_coupon_amount: {
-                type: DataTypes.DECIMAL,
+                type: DataTypes.FLOAT( 10, 2 ),
                 allowNull: true,
                 field: 'magic_coupon_amount',
                 defaultValue: 0.00
             },
             gst_rate: {
-                type: DataTypes.DECIMAL,
+                type: DataTypes.FLOAT( 10, 2 ),
                 allowNull: true,
                 field: 'gst_rate',
                 defaultValue: 0.00
             },
             gst_amt: {
-                type: DataTypes.DECIMAL,
+                type: DataTypes.FLOAT( 10, 2 ),
                 allowNull: true,
                 field: 'gst_amt',
                 defaultValue: 0.00
             },
             given_to_res: {
-                type: DataTypes.DECIMAL,
+                type: DataTypes.FLOAT( 10, 2 ),
                 allowNull: true,
                 field: 'given_to_res',
                 defaultValue: 0.00
             },
             discount_given: {
-                type: DataTypes.DECIMAL,
+                type: DataTypes.FLOAT( 10, 2 ),
                 allowNull: true,
                 field: 'discount_given',
                 defaultValue: 0.00
